requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
PyJWT==2.9.0
google-auth==2.35.0
python-dotenv==1.0.1
python-multipart==0.0.12